+++ src/consumer/component/consumer-component.js
@@ -102,6 +102,7 @@ export type ComponentProps = {
   files: SourceFile[],
   docs?: ?(Doclet[]),
   dists?: Dist[],
+  mainDistFile?: ?PathLinux,
   specsResults?: ?SpecsResults,
   license?: ?License,
   deprecated: ?boolean,
@@ -212,6 +213,7 @@ export default class Component {
     overrides,
     docs,
     dists,
+    mainDistFile,
     specsResults,
     license,
     log,
@@ -245,6 +247,7 @@ export default class Component {
     this.overrides = overrides;
     this._docs = docs;
     this.setDists(dists);
+    this.dists.setMainDistFile(mainDistFile ? path.normalize(mainDistFile) : null);
     this.specsResults = specsResults;
     this.license = license;
     this.log = log;
@@ -974,6 +977,7 @@ export default class Component {
     const deprecated = componentFromModel ? componentFromModel.deprecated : false;
     const componentDir = componentMap.getComponentDir();
     let dists = componentFromModel ? componentFromModel.dists.get() : undefined;
+    const mainDistFile = componentFromModel ? componentFromModel.dists.getMainDistFile() : undefined;
     const getLoadedFiles = async (): Promise<SourceFile[]> => {
       const sourceFiles = [];
       await componentMap.trackDirectoryChanges(consumer, id);
@@ -1103,6 +1107,7 @@ export default class Component {
       loadedFromFileSystem: true,
       componentMap,
       dists,
+      mainDistFile: mainDistFile ? path.normalize(mainDistFile) : null,
       compilerPackageDependencies,
       testerPackageDependencies,
       deprecated,